Backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cffi==1.17.1
celery==5.4.0
click==8.1.8
colorama==0.4.6
coverage==7.8.0
cryptography==44.0.2
dlib==19.24.8
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.12
greenlet==3.2.1
h11==0.16.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
passlib==1.7.4
pillow==11.2.1
pluggy==1.5.0
psycopg2-binary==2.9.10
redis==5.2.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pytest==8.3.5
pytest-cov==6.1.1
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
SQLAlchemy-Utils==0.41.2
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
